Preparing to remove detection_box_2d.gd
The plan is to just have a hit_box_2d script that is extended by a hurt_box_2d script. Both hit_box and hurt_box have identical features. The hurt box exist as a separate script only for showing a separate node / box color. Detection box is being removed as it is redundant.