Replace combat state machine with input advancer
Rather than having a specialized state machine with support for buffered inputs, this new node can be used to control the state machine and advance using buffered inputs. This change also removes the concept of "situations". If desired users can create sub-state-machines and use that to mimic the previous situation model.